Fully Functioning Orders and Menus,commit by danish

+++ app/controllers/menu_items_controller.rb
@@ -22,12 +22,16 @@ class MenuItemsController < ApplicationController
   def destroy
     id = params[:id]
     menu_item = MenuItem.find(id)
-    menu_item.destroy
     orders = Order.where(status: "notprocessed")
-    orders.each do |orders|
-      order_item = orders.order_items.find_by(menu_item_id: menu_item.id)
-      order_item.destroy
+    orders.each do |order|
+      if order.order_items
+        order_item = order.order_items.find_by(menu_item_id: menu_item.id)
+        if order_item
+          order_item.destroy
+        end
+      end
     end
+    menu_item.destroy
     redirect_to "/menus/#{session[:current_selected_menu_id]}"
   end
 end
